Move creating ChromiumContext to context.ts
